Segmentation: improved method `is_another_in_between` using intersection
Using line intersection algorithm to compute if a two symbols are
obstructed or not. This fixes the problem of direction, as before,
`is_another_in_between` can only find if an object is between only in
vertical positions (when symbols are on top of each other). But now this
will work with any position.