Parser: Used one way relations and some typing fixes
In `get_most_probable_relation` and `distance_labeled_crops` we are using
`Optional` type, meaning that it might return `None`. In our case it will
return `None` when we are trying to find the relation between two symbols but
the first is on the right of the second.

Also added some typing in some functions and declarations