Parsing: added method to return relations connections
We will need to know the relations strings for later when we build the tree,
that's why we need it.
In contrast, the function `get_all_symbols_normalized_connections` might not be
needed, as the normalized connection distance is only needed for the minimum
spanning tree graph logic. So we might make it private at some point.