Classification: added another special symbol `\\frac` for fraction
This symbol is not classified using the SVM model, but using heuristics and manual
checks done for the `minus(-)` symbol, we try to find if this `minus` has other
symbols on top and bottom of it, if it actually does, we make it `\\frac`.

Separating these two symbols makes it easier for us to identify relationships
between symbols.