Parsing: added recursive LaTeX from the symbol tree
We start from the root node, each node will generate the latex part for
itself and the children it has. We also have strict checks for the type
of children/relations a symbol might have. For example, the `\\frac`
cannot have `power` and `sub`.

For consistent output, added `optimize_latex_string` which would remove
whitespace as they are ignored by the latex compiler and also remove
the brackets `{}` if only one element is inside them, I think this is what
the latex compiler is doing from a simple check.

Now all core functions of the parser module is finished, we might need to
optimize the code more, and make it more robust in predicting
the symbol tree and the resulting LaTeX code.