Segmentation: fixed opencv cropping segments and readded flood-fill
The opencv cropping was using the old method of cropping the box, which is
wrong as sometimes adjacent characters might be in the cropped image.

Second, readded flood-fill implementation as it is the original plan of
using segmentation in this project. But will not be used in the main web
server.